requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
audioop-lts==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
croniter==5.0.1
dataclasses-json==0.6.7
discord-typings==0.9.0
discord.py==2.4.0
distro==1.9.0
dnspython==2.7.0
emoji==2.14.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pymongo==4.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.1.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.2
